Cutdown makefile install method
Since lunchbox is still beta and has no launcher
it's best to not exec it in your desktop session
since you won't be able to open anything.
You *could* open a virtual terminal, export
a display var and open programs that way but I
think that's annoying and virtual terminals
are disabled by default in many modern distress.

I've removed this footgun and the installation
rules. At this stage it's better if developers
compile and run it themselves from a local folder.