Fix incremental resizing again.
It is _very_ tricky to properly cooperate with windows that demand to have height or width increments (e.g., terminals, gimp toolbox) because if they aren't a size they like they will send us ConfigureRequests that we must listen to. This can result in a loop if we aren't careful that will result in crunching a window down to a minimum size. It's not just a matter of having the frame dimension be a multiple of a constant it must also consider another base aomunt that can be added. In the case of gnome-terminal, this is the menubar. (that is not a character height) I normally use the min_width, min_height hints to derive this information, but this doesn't always work. (e.g., the menubar is hidden on gnome-terminal) So I'm going to calculate this offset again to prevent this feedback loop.